app/requirements/requirements.txt
# app
fastapi==0.94.1
uvicorn==0.21.1
uvicorn[standard]
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.9

# db
sqlalchemy[asyncio]
asyncpg

# docx_parsing
aiofiles==23.2.1
regex==2024.5.15
python-docx==1.1.2